+++ Badge2022EF.Models/Concretes/Roles.cs
@@ -7,7 +7,7 @@
             Name = rolesName;
             //personnes = rolesPersonnes;
         }
-        public long Id { get; set; } = long.MinValue;
+        public long Id { get; set; } = int.MinValue;
         public string Name { get; set; } = string.Empty;
 
         public ICollection<Personnes>? Personnes { get; set; }
